All ready

+++ project-template/src/main/java/edu/hw4/Task18.java
@@ -5,7 +5,7 @@ import java.util.List;
 public class Task18 {
     @SafeVarargs public static Animal heaviestFish(List<Animal> ... listsAnimals){
         Animal heaviestFish = null;
-        int weightHeaviestFish = 0;
+        float weightHeaviestFish = 0;
         for(var listAnimals:listsAnimals){
             for(var animal:listAnimals){
                 if(animal.type() == Animal.Type.FISH && animal.weight() > weightHeaviestFish){
