+++ project-template/src/main/java/edu/hw4/Task8.java
@@ -5,7 +5,7 @@ import java.util.List;
 public class Task8 {
     public static Animal heaviestAnimalBelowFixedHeight(List<Animal> animals, int fixedHeight){
         Animal heaviestAnimal = null;
-        int weightHeaviestAnimal = 0;
+        float weightHeaviestAnimal = 0;
         for(var animal:animals){
             if(animal.height() < fixedHeight && weightHeaviestAnimal < animal.weight()){
                 weightHeaviestAnimal = animal.weight();
